Added HDR mode for primer design
This adds the changes we need for primer design for HDR. The HDR code path is activated by passing in a GET param 'hdrDist', which denotes the distance from the current guide to the HDR insert point. The primer design is centered on the insert point. 

The changes may be useful to other consumers in the current form,  or they may require proper integration into the UI, in which case this can remain an experimental or undocumented feature. 

NOTE: my editor removed all the extraneous whitespace automatically, resulting in many lines of change. See designPrimer for significant changes.